+++ packages/cli/commands/base/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @imohuan-plus/base-command
 
+## 1.0.1
+
+### Patch Changes
+
+- fix update
+
 ## 1.0.0
 
 ### Major Changes