+++ packages/cli/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @imohuan-plus/cli
 
+## 1.0.1
+
+### Patch Changes
+
+- fix update
+- Updated dependencies
+  - @imohuan-plus/npm-package@1.0.1
+
 ## 1.0.0
 
 ### Major Changes