+++ packages/play/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @imohuan-plus/play
 
+## 1.0.1
+
+### Patch Changes
+
+- fix update
+- Updated dependencies
+  - @imohuan-plus/components@1.0.1
+
 ## 1.0.0
 
 ### Major Changes