+++ packages/utils/log/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @imohuan-plus/log
 
+## 1.0.1
+
+### Patch Changes
+
+- fix update
+- Updated dependencies
+  - @imohuan-plus/utils-common@1.0.1
+
 ## 1.0.0
 
 ### Major Changes